Rewrite codebase (#14)
* Get rid of ESlint

* Update readme and scripts

* Many changes, part of #5, #6, and #7

1. Convert to NPM module
2. Bump and pin `discord.js` to 14.23.2
3. Bump and pin `dotenv` to 17.2.3
4. Remove `node-fetch` as Fetch is now formally supported in Node
5. Add `@types/node` as dev dependency
6. Rename all modules in `/modules` to .mjs
7. Enforce single quote style
8. Add `tsconfig.json` for Typescript type hinting
9. Gentrify `lommus.js`
10. Gentrify `deploy-commands.js`

* Add Discord chat icon to README

* Delete package-lock.json

* Suppress dotenv logging

* Wrap logic into `LoMMuS` class

This commit wraps the entirety of `lommus.js`'s logic into the `LoMMuS` class. This is the first step to completely modularizing the bot.

Also, enable TS strict null checks

* Normalize to LF EOL

* Add hybrid module loading #6

* ESM-ize Annoyance module

* Check loaded modules, move `module.mjs` to subdirectory

* ESM-ize Lomboard module

* Revert "ESM-ize Lomboard module"

This reverts commit 793e4e32a99863a5c40768710b511e4a4034e65e.

* Now properly ESM-ize Lomboard

* ESM-ize GitHub Issue Parser Module

* Expand README

* ESM-ize User Commands module

* Use `Ephemeral` flags

* Delete old LoMMuS script

* Rework module loading

* ESM-ize Role Selector Module

* ESM-ize Watchdog Module

* This ain't the right name

* Add `@types/lzma`

* ESM-ize Logging Module

* ESM-ize Moderation Module

* ESM-ize Role Selection Module

* Change ephemeral property to `Ephemeral` flag

* Use `globalThis` for global variables

The TODO still applies, this is just makeup on a rock

* Encapsulate `colorRandom` into `RoleSelectionModule`

* Broaden error guard

* Fix merge conflicts on rewrite branch (#15)

* Add ability to fetch issues from other repos (#9)

* Add ability to fetch issues from other repos, + add examples for .env and .config.json

* Update modules/githubIssueParser.js


* Add pm2 daemon scripts for Windows and Linux #13

* Add `pm2` daemon scripts for Windows and Linux.
Closes #13

---------


* Auto-normalize via gitattribute

* remove logging

---------

+++ modules/watchdog.mjs
@@ -1,17 +1,28 @@
-// const { blockQuote } = require('@discordjs/builders');
-const { EmbedBuilder, Events, PermissionFlagsBits } = require('discord.js');
-const { red } = require('../config.json');
+import { EmbedBuilder, Events, PermissionFlagsBits } from 'discord.js';
+import config from '../config.json' with { type: 'json' };
 // const badURLs = require('../spam.json');
-const wait = require('node:timers/promises').setTimeout;
-const urlReg = /https?:\/\/\w+/;
+import { setTimeout as wait } from 'node:timers/promises';
+import { BotModule } from './util/module.mjs';
 
-module.exports = {
+export default class WatchdogModule extends BotModule {
+	/**
+	 * Cache color configuration here + TS assertions
+	 * @constant
+	 */
+	colors = {
+		RED: /** @type {`#${string}`} */ (config.red)
+	};
 
-	name: 'Watchdog',
-	description: 'oh no you didn\'t.',
-	listeners: ['messageCreate'],
+	/**
+	 * The RegExp for URLs
+	 */
+	urlReg = RegExp(/https?:\/\/\w+/, "i");
 
-	async execute(client) {
+	constructor () {
+		super('Watchdog', "oh no you didn't.", ['messageCreate']);
+	}
+	/** @param {import('discord.js').Client} client */
+	init(client) {
 
 		/*	// Search Functions
 		const multiSearchOr = (text, searchWords) => (
@@ -48,12 +59,7 @@ module.exports = {
 			}
 		} */
 
-		/* ============================
-
-				Message Create
-
-		=============================*/
-		client.on(Events.MessageCreate, async message => {
+		client.on(Events.MessageCreate, async (message) => {
 			if (message.author.bot) return;
 
 			/* // Spam filtering
@@ -88,15 +94,20 @@ module.exports = {
 			} */
 
 			// resouces image/link handler
-			if (message.channel.name === 'resources') {
+			if ('name' in message.channel && message.channel.name === 'resources') {
 				// is it a staff? does it have no url and attachments?
-				if (message.member.permissions.has(PermissionFlagsBits.BanMembers) || !urlReg.test(message.content) && message.attachments.size > 0) {
+				if (
+					message.member
+					&& message.member.permissions.has(PermissionFlagsBits.BanMembers)
+					|| !this.urlReg.test(message.content)
+					&& message.attachments.size > 0
+				) {
 					message.react('✅');
 				}
 				else {
 					message.delete();
 					const embed = new EmbedBuilder()
-						.setColor(red)
+						.setColor(this.colors.RED)
 						.setDescription('This channel is for sharing presets, samples, and LMMS themes you have the legal rights to share. This is not a discussion or request channel. Please do not share binaries or links to external resources.');
 
 					message.channel.send({ embeds: [embed] })
@@ -106,7 +117,6 @@ module.exports = {
 						});
 				}
 			}
-
 		});
-	},
-};
+	}
+}
